# swagger.yml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: Generic Error response from server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/redjoker011/online-cafe/handlers
  Product:
    description: Product defines the structure of an API product
    properties:
      description:
        type: string
        x-go-name: Description
      id:
        description: |-
          The id for this product

          require: true
        format: int64
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
      price:
        format: float
        type: number
        x-go-name: Price
      sku:
        type: string
        x-go-name: SKU
    type: object
    x-go-package: github.com/redjoker011/online-cafe/data
  ValidationError:
    description: Validation Error
    properties:
      messages:
        items:
          type: string
        type: array
        x-go-name: Nessage
    type: object
    x-go-package: github.com/redjoker011/online-cafe/handlers
info:
  description: Documentation for Product API
  title: of Products API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns List of Products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: |-
        Create a new Product
        response:
        200: productsResponse
        422: errorValidation
        501: errorResponse
      operationId: createProduct
      tags:
      - products
  /products{id}:
    put:
      operationId: updateProduct
      parameters:
      - description: The ID of the product to delete from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    headers:
      Body:
        description: |-
          Error Message
          in: Body
    schema:
      $ref: '#/definitions/GenericError'
  noContent:
    description: No content is returned by this API endpoint
  productsResponse:
    description: A List of Products
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
  validationResponse:
    description: Validation error messages
    headers:
      Body:
        description: |-
          Collection of error messages
          in: Body
    schema:
      $ref: '#/definitions/ValidationError'
schemes:
- http
swagger: "2.0"